Move WorksIncludes into the display-common lib
