+++ data_api/snapshot_convertor/src/main/scala/uk/ac/wellcome/platform/snapshot_convertor/flow/ElasticsearchHitToDisplayWorkFlow.scala
@@ -4,8 +4,8 @@ import akka.NotUsed
 import akka.stream.scaladsl.Flow
 import com.twitter.inject.Logging
 import io.circe.generic.extras.JsonKey
-import uk.ac.wellcome.display.models.DisplayWork
-import uk.ac.wellcome.models.{AllWorksIncludes, IdentifiedWork}
+import uk.ac.wellcome.display.models.{AllWorksIncludes, DisplayWork}
+import uk.ac.wellcome.models.IdentifiedWork
 import uk.ac.wellcome.utils.JsonUtil._
 
 import scala.concurrent.ExecutionContext
