+++ sbt_common/display/src/main/scala/uk/ac/wellcome/display/models/DisplayWork.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.display.models
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.annotations.{ApiModel, ApiModelProperty}
-import uk.ac.wellcome.models.{IdentifiedWork, WorksIncludes}
+import uk.ac.wellcome.models.IdentifiedWork
 
 @ApiModel(
   value = "Work",
