+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/WorksIncludesTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models
+package uk.ac.wellcome.display.models
 
 import org.scalatest.{FunSpec, Matchers}
